MASSIVE UPDATE WITH BREAKING CHANGES!!!!
We are changing the apps focus from Collection first to Deck first.

This means large changes to useCards, and to the concept of Scope.

Basically scope is gone! It was really confusing and complicated. So I've decied that I'm just done with it.

The easier method is to use the URL to determine where a card should be sent to. It should almost always be added to a deck. This is how deck first works.

Each card will have additional settings to add cards to the collection (or both) instead. But that is currently set as a TODO.

For now any cards added on the deck page will be added to the deck. Cards added from the "Search Cards" button within the Navbar will be added to collection.

Next steps will be to add URL definitions to the Navbar search, and additional buttons to straight add to the colleciton.